+++ core-ui/src/components/Clusters/views/ClusterOverview/clusterDetailsNode.ts
@@ -1,4 +1,4 @@
-import { NavNode, Scope } from '../../types';
+import { NavNode, Scope } from '../../../../state/types';
 
 export const createClusterNode = (scope: Scope): NavNode => ({
   category: '',
