+++ core-ui/src/resources/ClusterEvents/index.js
@@ -1,8 +1,7 @@
-import { predefinedCategories } from 'state/navigation/categories';
-
 export { List, Details, resourceType } from 'resources/Events';
 export const namespaced = false;
 export const apiGroup = '';
 export const apiVersion = 'v1';
 
-export const category = predefinedCategories.temporary;
+export const category = '';
+export const topLevelNode = true;
